Decission Tree Practise Code
Decision trees are a type of supervised machine learning algorithm used for classification and regression tasks. They model decisions with tree-like graphs, where each node represents a feature, each branch represents a decision rule, and each leaf represents an outcome. The topmost node is called the root. The algorithm selects the best feature based on a certain criteria, like information gain or Gini index, and makes that the root. This process is repeated for the sub-nodes until the tree is built. Decision trees are simple to understand, visualize, and interpret, but can be prone to overfitting.